Have some news: But Greg is smart. And by the way, never wrestle him. You understand that? Never. Any guy that can do a body slam, he's my kind of ... He was my guy. I shouldn't say this, because -- there's nothing to be embarrassed about. So I was in Rome with a lot of the leaders from other countries talking about all sorts of things, and I heard about it. And we endorsed Greg very early, but I had heard that he body-slammed a reporter. "And he was way up. And he was way up. And I said, oh, this was like the day of the election, or just before, and I said, oh, this is terrible, he's going to lose the election. Then I said, well, wait a minute, I know Montana pretty well. I think it might help him. And it did!"
MOre news!!! But Greg is smart. And by the way, never wrestle him. You understand that? Never.
Any guy that can do a body slam, he's my kind of ... He was my guy.
I shouldn't say this, because -- there's nothing to be embarrassed about. So I was in Rome with a lot of the leaders from other countries talking about all sorts of things, and I heard about it. And we endorsed Greg very early, but I had heard that he body-slammed a reporter.
"And he was way up. And he was way up. And I said, oh, this was like the day of the election, or just before, and I said, oh, this is terrible, he's going to lose the election. Then I said, well, wait a minute, I know Montana pretty well. I think it might help him. And it did!"But Greg is smart. And by the way, never wrestle him. You understand that? Never.
Any guy that can do a body slam, he's my kind of ... He was my guy.
I shouldn't say this, because -- there's nothing to be embarrassed about. So I was in Rome with a lot of the leaders from other countries talking about all sorts of things, and I heard about it. And we endorsed Greg very early, but I had heard that he body-slammed a reporter.
"And he was way up. And he was way up. And I said, oh, this was like the day of the election, or just before, and I said, oh, this is terrible, he's going to lose the election. Then I said, well, wait a minute, I know Montana pretty well. I think it might help him. And it did!"